Pass request parameter from request handler to Remy
In Lwan, handlers are defined like this:

	function handle_METHOD_NAME(req)
	   req:say('Hello, World!')
	end

Where METHOD is the HTTP method ("get", "post", etc), and name is the
canonical prefix for that particular URI (a "GET /" will have the
special NAME of "root", while "/foobar?param=3" will have "foobar" as
its name).

The `req` parameter is the only way one communicates with the request,
either to obtain some values (such as the query parameters) or to send
a response.