Avoid autodetecting environment each and every request
This caches the result of remy.detect() inside the sailor table, and
reuses it next time remy.init() is called.  This works since the
`remy_mode` is `nil` the first time, which is a synonym for
`remy.MODE_AUTODETECT`; it is then assigned to whatever value defines
the current environment.