feat(page): Adds page:to_browser method
This method will take a variable on the server side, convert it to a string and wrap around client side Lua so it's accessible over there. Attention, it does not convert functions, only strings, tables and numbers. Currently it is only supported when using Starlight as the Lua-JS virtual machine.